Add new context class
`Context` represents the data needed for a command to be run. It is
meant to be both the data sent to commands, as well as the updated data
that is returned. Currently this data manipulation is happening in
place, but the current approach would allow us to easily switch to
copying the data so that we could keep a history of the changes if we
so desired.