`Computer` can now run a program
With this commit, the `Computer` class can now run a program which
consists of the PUSH, PRINT, and STOP commands. The main improvement
that could be done here would be to construct the command objects when
they are inserted into the stack to avoid needing to iterate over a
mostly empty stack.

Another caveat of this implementation is that if no STOP command is
present, this program will run forever.