Add new NULL instruction
The NULL instruction will be used for slots in the stack where no
instructions were inserted. The main reason for doing this is that we
can now easily customize, and test, the behavior of empty instructions.
If we want to raise, we can; if we want to do nothing (which we are
currently), we can.